backend/requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
colorama==0.4.6
coverage==7.2.7
cryptography==41.0.1
Django==4.0.4
django-cors-headers==3.14.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
linecache2==1.0.0
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
pycparser==2.21
PySocks==1.7.1
pytest==7.3.2
pytest-cov==4.1.0
pytz==2023.3
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
traceback2==1.4.0
trio==0.22.0
trio-websocket==0.10.3
tzdata==2023.3
unittest2==1.1.0
urllib3==2.0.3
wsproto==1.2.0
